commit 08/12/2566 11.21

+++ pages/api/itemPromotion/service.ts
@@ -1,5 +1,5 @@
 import { dataVerifyIUpdatetemPromotion, dataVerifyItemPromotion } from "@/types/verify";
-import { checkArrayPdIdInPromotionId, checkArrayProductId, checkArrayPromotionId, checkItemPromotionIdArr, checkItemPromotionProductIdArr, deleteDataItPromotionByPdIdPmId, deleteDataItemPromotion, fetchAllItemPromotion, fetchItemPromotionByCompanyId, fetchItemPromotionById, fetchItemPromotionByPromotionId, insertArrayPromotionItem, insertItemPromotion, updateDataItemPromotion, verifyItemPromotionBody } from "@/utils/itemPromotion";
+import { checkArrayPdIdInPromotionId, checkArrayProductId, checkArrayPromotionId, checkItemPromotionIdArr, checkItemPromotionProductIdArr, deleteDataItPromotionByPdIdPmId, deleteDataItemPromotion, fetchAllItemPromotion, fetchItemPromotionByCompanyId, fetchItemPromotionById, fetchItemPromotionByPromotionId, insertArrayPromotionItem, insertItemPromotion, updateDataItemPromotion, updateItemPromotionArr, verifyItemPromotionBody } from "@/utils/itemPromotion";
 import { NextApiResponse } from "next";
 
 export const handleAddItemPromotion = async (body: dataVerifyItemPromotion[], res: NextApiResponse) => {
@@ -24,28 +24,28 @@ export const handleAddItemPromotion = async (body: dataVerifyItemPromotion[], re
 
 export const handleGetItemPromotionById = async (res: NextApiResponse, id: number) => {
     const itemPromotion = await fetchItemPromotionById(id);
-    if (!itemPromotion) return res.status(404).json({ message: `No itemPromotion found with Id : ${id}`, itemPromotion: null, status: false });
+    if (!itemPromotion) return res.status(200).json({ message: `No itemPromotion found with Id : ${id}`, itemPromotion: null, status: false });
 
     return res.status(200).json({ message: "ItemPromotion found", itemPromotion: itemPromotion, status: true });
 }
 
 export const handleGetItemPromotionByPromotionId = async (res: NextApiResponse, promotionId: number) => {
     const itemPromotion = await fetchItemPromotionByPromotionId(promotionId);
-    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(404).json({ message: `No itemPromotion found with promotionId : ${promotionId}`, itemPromotion: null, status: false });
+    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(200).json({ message: `No itemPromotion found with promotionId : ${promotionId}`, itemPromotion: null, status: false });
 
     return res.status(200).json({ message: "itemPromotion found", itemPromotion: itemPromotion, status: true });
 }
 
 export const handleGetItemPromotionByCompanyId = async (res: NextApiResponse, companyId: number) => {
     const itemPromotion = await fetchItemPromotionByCompanyId(companyId);
-    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(404).json({ message: `No itemPromotion found with companyId : ${companyId}`, itemPromotion: null, status: false });
+    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(200).json({ message: `No itemPromotion found with companyId : ${companyId}`, itemPromotion: null, status: false });
 
     return res.status(200).json({ message: "itemPromotion found", itemPromotion: itemPromotion, status: true });
 }
 
 export const handleGetAllItemPromotion = async (res: NextApiResponse) => {
     const itemPromotion = await fetchAllItemPromotion();
-    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(404).json({ message: `No itemPromotion found`, itemPromotion: null, status: false });
+    if (!itemPromotion || (Array.isArray(itemPromotion) && itemPromotion.length === 0)) return res.status(200).json({ message: `No itemPromotion found`, itemPromotion: null, status: false });
 
     return res.status(200).json({ message: "itemPromotion found", itemPromotion: itemPromotion, status: true });
 }
@@ -65,10 +65,13 @@ export const handleUpdateItemPromotion = async (body: dataVerifyIUpdatetemPromot
     const checkProductId = await checkItemPromotionProductIdArr(body);
     if (checkProductId.length > 0) return res.status(404).json({ message: checkProductId, itemPromotion: null, status: false });
     // deleteItemPromotion By productId And promotionId
-    const deleteItemPromotion = await deleteDataItPromotionByPdIdPmId(body.deleteItemPromotionData);
+    if(body.deleteItemPromotionData.length > 0){
+        const deleteItemPromotion = await deleteDataItPromotionByPdIdPmId(body.deleteItemPromotionData);
+        if (deleteItemPromotion.length > 0) return res.status(404).json({ message: deleteItemPromotion, itemPromotion: null, status: false });
+    }
     // updateItemPromotion
-    // const updateItemPromotion = await updateDataItemPromotion(body, body.id);
-    // if (!updateItemPromotion) return res.status(404).json({ message: "An error occurred saving data.", itemPromotion: null, status: false });
+    const updateItemPromotion = await updateItemPromotionArr(body.itemPromotionData);
+    if (updateItemPromotion.length > 0) return res.status(404).json({ message: updateItemPromotion, itemPromotion: null, status: false });
 
     return res.status(200).json({ message: "Data saved successfully.", itemPromotion: [], status: true });
 }
